Separate manifest parser contracts
